requirements.txt
agno
mcp
openai
python-dotenv
requests
pytz
crewai
ollama
langgraph 
langchain 
litellm 
langchain-community 
langchain_openai
geopy
langchain-ollama
autogen
pymupdf
pypdf
streamlit
faiss-cpu
deepeval